requirements.txt

langchain
langgraph
tavily-search
openai
tiktoken
faiss-cpu
python-dotenv
